Logic, comment, and feature updates
Comment remaining methods
Notify user with toasts
Add a maximum to the coffee order
Add a method to reset our state after an order is placed
Add a method to execute orders to get the complexity out of the button click event.
Use a string formatter wherever possible to make sure we use local formatting rules.